Add `buy` command
### Probably one of the last core commits to the new currency API
It is the buy command. Probably the most useful! You can specify an item to buy, along with it's quantity, if you want to. The default quantity is set to 1.
There is currently no restriction on how many you can buy at one shot, but it will depend on your wallet balance. ALSO, the quantity cannot be below 1.