Fix duplicate naming error causing save fails when `save()` is called
The conflict was caused due to the `/open` command, which is also used as another command in the `save()` command. Due to this, whenever the database was supposed to save cached data, it would fail and return `BaseCommandObject` exception instead.

This patch should fix it, and allow the database to save data correctly.