Fix a bug which returns `KeyError` when modifying user level
When the levelling db library attempts to add or remove levels from a user (through framework), it ends up failing and returning a `KeyError` exception.

This was because in the levelling db library, the `add_levels()` and `remove_levels()` methods were attempting to modify the `levels` value (non-existant) instead of the `level` value (actual key name).

This patch aims at resolving this issue.