Add isobot server activity logger
This feature will allow the bot to show how active a specific guild is. This also has support for new DM flagging.

For privacy reasons, any message or media content is not logged to the bot client. Only the username, server name and channel name are logged.

This is currently a basic implementation of the feature, actual commands will come later in the future for everyone to enjoy!